# `DateTime::to_rfc_*` optimizations