--- a/src/format/formatting.rs
+++ b/src/format/formatting.rs
@@ -528,9 +528,42 @@ pub(crate) fn write_rfc3339(
     dt: NaiveDateTime,
     off: FixedOffset,
 ) -> fmt::Result {
-    // reuse `Debug` impls which already print ISO 8601 format.
-    // this is faster in this way.
-    write!(w, "{:?}", dt)?;
+    let year = dt.date().year();
+    if (0..=9999).contains(&year) {
+        write_hundreds(w, (year / 100) as u8)?;
+        write_hundreds(w, (year % 100) as u8)?;
+    } else {
+        // ISO 8601 requires the explicit sign for out-of-range years
+        write!(w, "{:+05}", year)?;
+    }
+    w.write_char('-')?;
+    write_hundreds(w, dt.date().month() as u8)?;
+    w.write_char('-')?;
+    write_hundreds(w, dt.date().day() as u8)?;
+
+    w.write_char('T')?;
+
+    let (hour, min, mut sec) = dt.time().hms();
+    let mut nano = dt.nanosecond();
+    if nano >= 1_000_000_000 {
+        sec += 1;
+        nano -= 1_000_000_000;
+    }
+    write_hundreds(w, hour as u8)?;
+    w.write_char(':')?;
+    write_hundreds(w, min as u8)?;
+    w.write_char(':')?;
+    let sec = sec;
+    write_hundreds(w, sec as u8)?;
+
+    if nano == 0 {
+    } else if nano % 1_000_000 == 0 {
+        write!(w, ".{:03}", nano / 1_000_000)?;
+    } else if nano % 1_000 == 0 {
+        write!(w, ".{:06}", nano / 1_000)?;
+    } else {
+        write!(w, ".{:09}", nano)?;
+    }
 
     OffsetFormat {
         precision: OffsetPrecision::Minutes,