--- a/src/format/formatting.rs
+++ b/src/format/formatting.rs
@@ -15,7 +15,7 @@ use crate::naive::{NaiveDate, NaiveDateTime, NaiveTime};
 #[cfg(any(feature = "alloc", feature = "std"))]
 use crate::offset::{FixedOffset, Offset};
 #[cfg(any(feature = "alloc", feature = "std"))]
-use crate::{Datelike, Timelike, Weekday};
+use crate::{Datelike, SecondsFormat, Timelike, Weekday};
 
 use super::locales;
 #[cfg(any(feature = "alloc", feature = "std"))]
@@ -427,7 +427,13 @@ fn format_inner(
                 // same as `%Y-%m-%dT%H:%M:%S%.f%:z`
                 {
                     if let (Some(d), Some(t), Some(&(_, off))) = (date, time, off) {
-                        Some(write_rfc3339(w, NaiveDateTime::new(*d, *t), off))
+                        Some(write_rfc3339(
+                            w,
+                            crate::NaiveDateTime::new(*d, *t),
+                            off.fix(),
+                            SecondsFormat::AutoSi,
+                            false,
+                        ))
                     } else {
                         None
                     }
@@ -523,10 +529,13 @@ impl OffsetFormat {
 
 /// Writes the date, time and offset to the string. same as `%Y-%m-%dT%H:%M:%S%.f%:z`
 #[cfg(any(feature = "alloc", feature = "std"))]
+#[inline]
 pub(crate) fn write_rfc3339(
     w: &mut impl Write,
     dt: NaiveDateTime,
     off: FixedOffset,
+    secform: SecondsFormat,
+    use_z: bool,
 ) -> fmt::Result {
     let year = dt.date().year();
     if (0..=9999).contains(&year) {
@@ -556,19 +565,28 @@ pub(crate) fn write_rfc3339(
     let sec = sec;
     write_hundreds(w, sec as u8)?;
 
-    if nano == 0 {
-    } else if nano % 1_000_000 == 0 {
-        write!(w, ".{:03}", nano / 1_000_000)?;
-    } else if nano % 1_000 == 0 {
-        write!(w, ".{:06}", nano / 1_000)?;
-    } else {
-        write!(w, ".{:09}", nano)?;
-    }
+    match secform {
+        SecondsFormat::Secs => {}
+        SecondsFormat::Millis => write!(w, ".{:03}", nano / 1_000_000)?,
+        SecondsFormat::Micros => write!(w, ".{:06}", nano / 1000)?,
+        SecondsFormat::Nanos => write!(w, ".{:09}", nano)?,
+        SecondsFormat::AutoSi => {
+            if nano == 0 {
+            } else if nano % 1_000_000 == 0 {
+                write!(w, ".{:03}", nano / 1_000_000)?
+            } else if nano % 1_000 == 0 {
+                write!(w, ".{:06}", nano / 1_000)?
+            } else {
+                write!(w, ".{:09}", nano)?
+            }
+        }
+        SecondsFormat::__NonExhaustive => unreachable!(),
+    };
 
     OffsetFormat {
         precision: OffsetPrecision::Minutes,
         colons: Colons::Colon,
-        allow_zulu: false,
+        allow_zulu: use_z,
         padding: Pad::Zero,
     }
     .format(w, off)