Split out of https://github.com/chronotope/chrono/pull/1196.

Previously `to_rfc3339` reused the `Debug` implementation of `DateTime`, which in turn used those of `FixedOffset`, `NaiveDateTime`, `NaiveDate` and `NaiveTime`. Going through the formatting machinery of the standard library for each adds some overhead. Writing it out manually is only 35 lines.

Having everything in one function also makes it easy to merge the completely different implementation in `to_rfc3339_opt`.

Before:
```
bench_datetime_to_rfc2822
                        time:   [98.849 ns 99.212 ns 99.611 ns]
Found 9 outliers among 100 measurements (9.00%)
  8 (8.00%) high mild
  1 (1.00%) high severe

bench_datetime_to_rfc3339
                        time:   [198.38 ns 199.69 ns 201.17 ns]
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

bench_datetime_to_rfc3339_opts
                        time:   [702.91 ns 705.77 ns 708.85 ns]
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
```

After:
```
bench_datetime_to_rfc2822
                        time:   [100.66 ns 101.16 ns 101.73 ns]
                        change: [+0.5475% +1.2171% +1.8937%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

bench_datetime_to_rfc3339
                        time:   [128.61 ns 130.09 ns 131.72 ns]
                        change: [-35.653% -34.787% -33.971%] (p = 0.00 < 0.05)
                        Performance has improved.

bench_datetime_to_rfc3339_opts
                        time:   [125.60 ns 126.14 ns 126.79 ns]
                        change: [-82.265% -82.158% -82.037%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
```

`DateTime::to_rfc3339_opts` sees the best improvement from 700 ns down to 130 ns.